Revised per @Umofomia feedback
The gotcha is that the previous version used "origin" which is a shortcut
for the local copy of the default branch. So, it would be a shortcut
to "origin/master". If new commits come in and you run a "fetch", it means
that the default branch is now different. I missed this b/c in my testing,
I hadn't done a fetch, so it seemed correct.